requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-sslserver>=0.22
gunicorn
whitenoise
django-environ
psycopg2-binary
reportlab
pillow
